requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.2
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.8
lxml==4.4.2
Markdown==3.1.1
mysqlclient==1.4.6
numpy==1.18.1
pandas==0.25.3
parse==1.14.0
psycopg2==2.8.4
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.1
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.2
sqlparse==0.3.0
tqdm==4.41.1
urllib3==1.25.3
w3lib==1.21.0
websockets==8.1
whitenoise==5.0.1
